+++ src/install.js
@@ -1,4 +1,0 @@
-import sagaPlugin from './saga'
-import { activatePlugin } from 'kea'
-
-activatePlugin(sagaPlugin)