+++ src/saga/index.js
@@ -6,7 +6,7 @@ import { keaSaga, startSaga, cancelSaga } from './saga'
 import { createSaga } from './create-saga'
 import { select } from 'redux-saga/effects'
 
-export default ({ useLegacyUnboundActions = true } = {}) => ({
+export default ({ useLegacyUnboundActions = false } = {}) => ({
   name: 'saga',
 
   defaults: () => ({